+++ JKClient/Handlers/JO/JONetHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace JKClient {
 	public abstract class JONetHandler : NetHandler {
 		public override int MaxMessageLength => 16384;
-		public JONetHandler(ProtocolVersion protocol) : base(protocol) {}
+		public JONetHandler(ProtocolVersion protocol) : base((int)protocol) {}
 	}
 }
